added projectService Util for project search
